Revert "Refactor app.py to use centralized models module"
