--- models/__init__.py
@@ -1,6 +1,0 @@
-from flask_sqlalchemy import SQLAlchemy
-db = SQLAlchemy()
-
-from .user import User
-from .search_history import SearchHistory
-from .ip_report import IPReport
