--- models/ip_report.py
@@ -1,12 +1,0 @@
-from . import db
-from datetime import datetime, timezone
-
-class IPReport(db.Model):
-    id = db.Column(db.Integer, primary_key=True)
-    ip_address = db.Column(db.String(45), nullable=False)
-    user_id = db.Column(db.Integer, db.ForeignKey('user.id'), nullable=False)
-    report_type = db.Column(db.String(50), nullable=False)  # 'malicious', 'spam', 'suspicious', etc.
-    comment = db.Column(db.Text, nullable=False)
-    timestamp = db.Column(db.DateTime, default=datetime.now(timezone.utc))
-
-    user = db.relationship('User', backref=db.backref('reports', lazy=True))
