+++ services/__init__.py
@@ -5,4 +5,3 @@ from .abuse import *
 from .url_shortener import *
 from .tor_lookup import *
 from .mail_checker import *
-from .rate_limit import *
