+++ services/vpn_proxy.py
@@ -1,5 +1,4 @@
 import requests
-from .ip_list import *
 
 def lookup_proxycheck(ip_address):
     """Check VPN/Proxy using proxycheck.io (free tier - 100 queries/day)"""
@@ -346,7 +345,7 @@ def get_vpn_provider_info(ip_address):
 
 def check_apple_nordvpn(ip_address):
     # Ensure lists are loaded
-    apple_ip_ranges, nordvpn_ips = download_and_parse_ip_lists()
+    download_and_parse_ip_lists()
     ip = ipaddress.ip_address(ip_address)
     # Check Apple
     for net in apple_ip_ranges:
